logging stuff

+++ modules/monitoring/app/de/objectcode/play2/plugin/monitoring/infoadapter/DbInfoAdapter.java
@@ -1,7 +1,7 @@
 package de.objectcode.play2.plugin.monitoring.infoadapter;
 
 public interface DbInfoAdapter {
-	public long getCreatedConnectionCount();
-	public long getFreeConnectionCount();
-	public long getLeasedCounnectionCount();
+	public int getCreatedConnectionCount();
+	public int getFreeConnectionCount();
+	public int getLeasedCounnectionCount();
 }
